modularized process_command for norm, created redirect_child.c to hold functions for child redirections, created redundant_utils.c for potentially redundant functions
