+++ include/execution.h
@@ -6,7 +6,7 @@
 /*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 16:53:36 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/12 13:34:19 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/12 14:19:15 by jteissie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,10 +32,12 @@ void	handle_error(char *message, int code);
 void	try_absolute_path(char **command, char **env);
 void	execute(char *av, char **env);
 
-int		open_files(int file_fd[], char *outfile, char *infile);
+int		open_files(int file_fd[], t_lex_parser *table);
+void	get_redirections(t_lex_parser *table, char *redirection[]);
 void	go_to_first_table(t_lex_parser *roaming, t_lex_parser *parsed);
 
 int		execute_commands(t_lex_parser *tables, char **envp);
 void	process_command(t_lex_parser *parsed, char **envp);
+void	redirect_child(int file_fd[], int p_fd[]);
 
 #endif
