+++ src/execution/execution_utils.c
@@ -6,42 +6,51 @@
 /*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 18:14:25 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/12 13:44:59 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/12 14:18:21 by jteissie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	handle_error(char *message, int code)
+void	get_redirections(t_lex_parser *table, char *redirection[])
 {
-	ft_putstr_fd(message, 2);
-	ft_putchar_fd('\n', 2);
-	exit(code);
-}
+	t_lex_parser		*roaming;
+	t_redirect_table	*redir;
+	char				*outfile;
+	char				*infile;
 
-void	trash(char **array)
-{
-	int	i;
-
-	i = 0;
-	if (!array)
-		return ;
-	while (array[i])
+	roaming = NULL;
+	infile = NULL;
+	outfile = NULL;
+	go_to_first_table(roaming, table);
+	while (roaming && roaming->type != TK_PARS_PIPE)
 	{
-		free(array[i]);
-		i++;
+		if (roaming->type == TK_PARS_REDIR)
+		{
+			redir = roaming->table;
+			if (redir->type == TK_PARS_IN)
+				infile = redir->redir_str;
+			else if (redir->type == TK_PARS_OUT)
+				outfile = redir->redir_str;
+			roaming->type = TK_PARS_RESERVED;
+		}
+		roaming = roaming->next;
 	}
-	free(array);
+	redirection[0] = infile;
+	redirection[1] = outfile;
 }
 
-int	open_files(int file_fd[], char *infile, char *outfile)
+int	open_files(int file_fd[], t_lex_parser *table)
 {
+	char	*redir[2];
+
 	file_fd[0] = 0;
 	file_fd[1] = 0;
-	if (infile)
-		file_fd[0] = open(infile, O_CREAT | O_TRUNC, 064);
-	if (outfile)
-		file_fd[1] = open(outfile, O_CREAT | O_TRUNC, 064);
+	get_redirections(table, redir);
+	if (redir[0])
+		file_fd[0] = open(redir[0], O_CREAT | O_TRUNC, 064);
+	if (redir[1])
+		file_fd[1] = open(redir[1], O_CREAT | O_TRUNC, 064);
 	if (file_fd[0] < 0 || file_fd[1] < 0)
 		return (-1);
 	return (SUCCESS);
@@ -52,30 +61,4 @@ void	go_to_first_table(t_lex_parser *roaming, t_lex_parser *parsed)
 	roaming = parsed;
 	while (roaming->prev && roaming->prev->type != TK_PARS_PIPE)
 		roaming = roaming->prev;
-}
-
-char	*ft_str_realloc_join(char *stash, char *add)
-{
-	char	*joined;
-
-	if (!stash)
-		return (NULL);
-	joined = ft_calloc((ft_strlen(stash) + ft_strlen(add) + 1), sizeof(char));
-	if (!joined)
-		return (free(stash), NULL);
-	copycat_str(joined, stash, add);
-	return (free(stash), joined);
-}
-
-void	copycat_str(char *out, char *cpy_src, char *cat_src)
-{
-	size_t	i;
-	int		src_i;
-
-	src_i = 0;
-	ft_strlcpy(out, cpy_src, ft_strlen(cpy_src) + 1);
-	i = ft_strlen(out);
-	while (cat_src[src_i])
-		out[i++] = cat_src[src_i++];
-	out[i] = '\0';
 }
