+++ src/execution/redundant_utils.c
@@ -1,0 +1,61 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   redundant_utils.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/07/12 13:53:38 by jteissie          #+#    #+#             */
+/*   Updated: 2024/07/12 13:54:09 by jteissie         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	handle_error(char *message, int code)
+{
+	ft_putstr_fd(message, 2);
+	ft_putchar_fd('\n', 2);
+	exit(code);
+}
+
+void	trash(char **array)
+{
+	int	i;
+
+	i = 0;
+	if (!array)
+		return ;
+	while (array[i])
+	{
+		free(array[i]);
+		i++;
+	}
+	free(array);
+}
+
+char	*ft_str_realloc_join(char *stash, char *add)
+{
+	char	*joined;
+
+	if (!stash)
+		return (NULL);
+	joined = ft_calloc((ft_strlen(stash) + ft_strlen(add) + 1), sizeof(char));
+	if (!joined)
+		return (free(stash), NULL);
+	copycat_str(joined, stash, add);
+	return (free(stash), joined);
+}
+
+void	copycat_str(char *out, char *cpy_src, char *cat_src)
+{
+	size_t	i;
+	int		src_i;
+
+	src_i = 0;
+	ft_strlcpy(out, cpy_src, ft_strlen(cpy_src) + 1);
+	i = ft_strlen(out);
+	while (cat_src[src_i])
+		out[i++] = cat_src[src_i++];
+	out[i] = '\0';
+}
